Apply combos to key_buffer.
Combos are now applied to the key_buffer from which they are then
executed just like normal keys in the buffer.

This means that multiple combos can be executed at the same time.

The old "timer is reset on every key press" is back as the default
behaviour but a new option `COMBO_STRICT_TIMER` sets the behaviour to
setting the timer only on the first key press.

COMBO_PERMISSIVE_HOLD is gone as it is now the "default". This is a side
effect of having combos execute at the same time.